Adding Support for IAM-based Authorization.
This patch adds support to dynamically generate passwords
necessary to access Redshift by using the Redshfit API
GetClusterCredentials().

https://docs.aws.amazon.com/redshift/latest/APIReference/API_GetClusterCredentials.html

It intercepts the extraction of the connection arguments, if the
new additional `iam` parameter is present. It expects the
hostname parameter of the URL to be the actual cluster identifier
and will use this to lookup the DNS name and port configuration.
In addition, it updates the username and password with the
updated values retrieved from the API call.

It expects the URL for the engine to be of the following structure:

    redshift://user@cluster_identifier/database?iam=true

Testing:

This patch adds two new unit tests for parsing the URL and extracting
the correct parameters from the API calls. It adds boto3 to the Tox
dependencies for mock testing.